Move create_basics, create_asset to ddo_helpers
